Moved functions in __init__ to a frontend module
